restructure: move common/__init__.py to ui/common.py, to allow launchpad builds.
Launchpad builds import common which tries to import gi which isn't available.
