+++ epoptes/ui/exec_command.py
@@ -4,7 +4,8 @@
 """
 Execute command dialog.
 """
-from epoptes.common import config, locate_resource
+from epoptes.common import config
+from epoptes.ui.common import locate_resource
 from gi.repository import Gtk
